Add files via upload
The main page for the website, it utilizes skeleton.css for responsiveness and layout. The links on top will go ta new page with some trivial information regarding the food selected. On smaller screens a dropdown will appear in it's place but does the same thing when selected. The form submits information using ajax to server.php which utilizes the users.php library to do the parsing. It then updates the page with the results of the parsing.