Add files via upload
food.php is a very simple webpage that displays trivial information about the food item link that was selected. The actual information it displays is stored in the foods.php library, so it calls this library with whatever food item was selected in the request, then uses PHP to drop the information into a boilerplate layout.